UPSTREAM: net: Clone skb before setting peeked flag
Shared skbs must not be modified and this is crucial for broadcast
and/or multicast paths where we use it as an optimisation to avoid
unnecessary cloning.

The function skb_recv_datagram breaks this rule by setting peeked
without cloning the skb first.  This causes funky races which leads
to double-free.

This patch fixes this by cloning the skb and replacing the skb
in the list when setting skb->peeked.

Fixes: a59322be07c9 ("[UDP]: Only increment counter on first peek/recv")
Reported-by: Konstantin Khlebnikov <khlebnikov@yandex-team.ru>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 738ac1ebb96d02e0d23bc320302a6ea94c612dec)

BUG=chrome-os-partner:52665
TEST=Run suspend_stress_test on oak/elm and make sure there's no
     memory corruption.

Change-Id: Ib1ceba30a33765a44cf46eb9e03d609b72d8349d
Signed-off-by: Ricky Liang <jcliang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/344623
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>