UPSTREAM: arm64: head.S: initialise mdcr_el2 in el2_setup
When entering the kernel at EL2, we fail to initialise the MDCR_EL2
register which controls debug access and PMU capabilities at EL1.

This patch ensures that the register is initialised so that all traps
are disabled and all the PMU counters are available to the host. When a
guest is scheduled, KVM takes care to configure trapping appropriately.

Cc: <stable@vger.kernel.org>
Acked-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
(cherry picked from commit d10bcd473301888f957ec4b6b12aa3621be78d59)

BUG=chrome-os-partner:54420
TEST=Run `perf -ad sleep 3` on my oak and make sure the counters
     values are not corrupted.

Change-Id: Ia5cff75b88d96c553b07a3b4b06ec043c7af04f9
Signed-off-by: Ricky Liang <jcliang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/353951
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>