CHROMIUM: drm/nouveau: Add delta to nvkm_vm_get_*()
Add delta to nvkm_vm_get() and nvkm_vm_get_offset().  We then set the delta
and length of the vma to track the exact (sub)region of the buffer that this
vma will represent.  This is to support the partial mapping/unmapping of a
buffer.

BUG=chrome-os-partner:49979
TEST=Build and boot to desktop on A44

Change-Id: Ib8daa70775ced0c14e3210f0491245749bae6973
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/329875
Reviewed-by: Vince Hsu <vince.h@nvidia.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/339285
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Stephen Barber <smbarber@chromium.org>
Tested-by: Stephen Barber <smbarber@chromium.org>