CHROMIUM: drm/nouveau: Fix map/unmap ioctls
Add tile_mode and tile_flags to gem map ioctl, so that we can specify
different tile modes and flags on a per mapping basis, which is a
requirement for supporting vulkan.  Also add offset to unmap to more
specifically identify the mapping to unmap, since we are allowed to map
the same buffer region to multiple GPU addresses.

BUG=chrome-os-partner:49979
TEST=Build and boot to desktop on A44.  Ran NVIDIA "MosaicMapping" tests
to verify that partial mappings work for both big and small pages.

Change-Id: Id2d853622b3a1f90611e10ef6d0f61950d089c3e
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/330884
Tested-by: Sean Paul <seanpaul@google.com>
Reviewed-by: Sean Paul <seanpaul@google.com>
Reviewed-on: https://chromium-review.googlesource.com/339303
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Tested-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Stephen Barber <smbarber@chromium.org>