CHROMIUM: drm/nouveau: Clear VOL bit when freeing AS
When we unmap a sparse buffer, we don't really unmap it, instead
we reset the VOL bit of it's PTE so that it can be reused later.
But if we're going to free the AS, we need to set PTE VOL bit to
zero otherwise GPU will think it's a valid PTE for sparse buffer.

BUG=chrome-os-partner:49979
TEST=Pass following two vulkan tests:
dEQP-VK.api.smoke.asm_triangle
dEQP-VK.api.device_init.create_device

Change-Id: Icef797cafb7111f6e14246feff2166683d23ae32
Signed-off-by: Shih-hsin Li <seasonl@nvidia.com>
Signed-off-by: Mark Zhang <markz@nvidia.com>
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/334621
Reviewed-on: https://chromium-review.googlesource.com/339309
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Tested-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Stephen Barber <smbarber@chromium.org>